BMC-219: Use the `appState` to show launching while loading the study
For whatever reason, iOS 14.4 is not recognizing the changes to the published Study
object. This is the only case where a Kotlin model object is used referenced directly
within SwiftUI. This changes the updateAppState() method to use that to set the
`appState` to `.launching` if the Study is being loaded.